Restore ComputingAndNetworkModificationType type as before and solve cyclic dependencies
