+++ src/features/parameters/loadflow/load-flow-parameters-inline.tsx
@@ -12,12 +12,11 @@ import type { UUID } from 'node:crypto';
 import { LoadFlowProvider } from './load-flow-parameters-provider';
 import { parametersStyles } from '../parameters-style';
 import { UseParametersBackendReturnProps } from '../../../utils/types/parameters.type';
-import { ComputingType } from '../common/computing-type';
 import { TreeViewFinderNodeProps } from '../../../components/ui/treeViewFinder';
 import { useSnackMessage } from '../../../hooks';
 import { SubmitButton } from '../../../components/ui';
 import { LabelledButton } from '../common/parameters';
-import { ElementType, GsLang, mergeSx } from '../../../utils';
+import { ComputingType, ElementType, GsLang, mergeSx } from '../../../utils';
 import { LineSeparator } from '../common';
 import { DirectoryItemSelector } from '../../../components/ui/directoryItemSelector';
 import { fetchLoadFlowParameters } from '../../../services/loadflow';
