+++ src/features/parameters/security-analysis/security-analysis-parameters-dialog.tsx
@@ -6,7 +6,7 @@
  */
 import { OptionalServicesStatus, useParametersBackend } from '../../../hooks';
 import { useSecurityAnalysisParametersForm } from './use-security-analysis-parameters-form';
-import { ComputingType, ParametersEditionDialogProps } from '../common';
+import { ParametersEditionDialogProps } from '../common';
 import {
     fetchSecurityAnalysisParameters,
     fetchSecurityAnalysisProviders,
@@ -15,7 +15,7 @@ import {
 } from '../../../services/security-analysis';
 import { SecurityAnalysisParametersForm } from './security-analysis-parameters-form';
 import { NameElementEditorForm } from '../common/name-element-editor';
-import { ElementType } from '../../../utils';
+import { ComputingType, ElementType } from '../../../utils';
 import { CustomMuiDialog } from '../../../components/ui/dialogs';
 
 export function SecurityAnalysisParametersDialog({
