+++ src/features/parameters/sensi/sensitivity-analysis-parameters-dialog.tsx
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-import { ComputingType, ParametersEditionDialogProps } from '../common';
+import { ParametersEditionDialogProps } from '../common';
 import { OptionalServicesStatus, useParametersBackend } from '../../../hooks';
 import {
     fetchSensitivityAnalysisParameters,
@@ -13,7 +13,7 @@ import {
 } from '../../../services/sensitivity-analysis';
 import { CustomMuiDialog } from '../../../components/ui/dialogs';
 import { NameElementEditorForm } from '../common/name-element-editor';
-import { ElementType } from '../../../utils';
+import { ComputingType, ElementType } from '../../../utils';
 import { useSensitivityAnalysisParametersForm } from './use-sensitivity-analysis-parameters';
 import { SensitivityAnalysisParametersForm } from './sensitivity-analysis-parameters-form';
