+++ src/features/parameters/short-circuit/short-circuit-parameters-inline.tsx
@@ -12,8 +12,8 @@ import type { UUID } from 'node:crypto';
 import { TreeViewFinderNodeProps } from '../../../components/ui/treeViewFinder';
 import { useSnackMessage } from '../../../hooks';
 import { SubmitButton } from '../../../components/ui';
-import { ElementType, UseParametersBackendReturnProps } from '../../../utils';
-import { ComputingType, LabelledButton } from '../common';
+import { ComputingType, ElementType, UseParametersBackendReturnProps } from '../../../utils';
+import { LabelledButton } from '../common';
 import { DirectoryItemSelector } from '../../../components/ui/directoryItemSelector';
 import { CreateParameterDialog } from '../common/parameters-creation-dialog';
 import { ShortCircuitParametersInfos } from './short-circuit-parameters.type';
