+++ src/utils/types/index.ts
@@ -24,3 +24,4 @@ export * from './dynamic-margin-calculation.type';
 export * from './network-modification-types';
 export * from './pcc-min.type';
 export * from './network-modification-metadata';
+export { ComputingType } from './computing-type';
